requirements.txt
Flask==2.3.3
requests==2.31.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
beautifulsoup4==4.12.2
feedparser==6.0.10
plotly==5.16.1
SQLAlchemy==2.0.20
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
nltk==3.8.1
vaderSentiment==3.3.2
praw==7.7.1